+++ lib/src/presentation/features/authentication/login/riverpod/login_provider.dart
@@ -2,7 +2,7 @@ import 'package:flutter_template/src/core/base/result.dart';
 import 'package:flutter_template/src/core/di/dependency_injection.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
-import '../../../../../domain/use_cases/authentication_use_cases.dart';
+import '../../../../../domain/use_cases/authentication_use_case.dart';
 import 'login_state.dart';
 
 part 'login_provider.g.dart';
